Refactor and relocate account assertion methods in tests
The methods assert_accounts_by_owner and assert_account_proofs_for_photon_and_test_indexer are moved from interop_nullify_test.rs to test_utils.rs.
